Revert "mac80211: QoS related cleanups"
This reverts commit f54ce8d94437bbbef417f9f57c54d3b66117499b.

+++ drivers/net/wireless/b43legacy/main.c
@@ -2383,7 +2383,8 @@ out:
 	return NETDEV_TX_OK;
 }
 
-static int b43legacy_op_conf_tx(struct ieee80211_hw *hw, u16 queue,
+static int b43legacy_op_conf_tx(struct ieee80211_hw *hw,
+				int queue,
 				const struct ieee80211_tx_queue_params *params)
 {
 	return 0;
